+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ This document describes changes between each past release.
 7.6.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**Bug fixes**
+
+- Fix removing permissions with Json Merge (#1322).
 
 
 7.5.1 (2017-10-03)